refactor: remove machine suffix from file names
