--- packages/element/src/viewer-context.ts
@@ -1,5 +1,0 @@
-import { createContext } from '@lit/context';
-import type { Viewer } from '@itk-viewer/viewer/viewer.js';
-export type { Viewer } from '@itk-viewer/viewer/viewer.js';
-
-export const viewerContext = createContext<Viewer>(Symbol('viewer'));
