+++ packages/viewer/src/fps-watcher.ts
